requirements.txt
pymupdf==1.24.2
langchain==0.1.17
langchain-community==0.0.36
faiss-cpu==1.8.0
huggingface-hub==0.21.4
transformers==4.38.2
sentence-transformers==2.3.1
streamlit==1.31.1
python-dotenv==1.0.0
openai==0.28
supabase==2.3.0
